Added plant_disease model
