+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "tinyfl/models/plant_disease_model"]
+	path = tinyfl/models/plant_disease_model
+	url = https://github.com/Ceres445/plant-disease-classification-pytorch
