+++ configs/agg.config.json
@@ -1,7 +1,7 @@
 {
   "host": "127.0.0.1",
   "port": 5000,
-  "consensus": 2,
+  "consensus": 1,
   "timeout": 1000,
   "epochs": 3,
   "strategy": "fedavg"
