+++ tinyfl/model.py
@@ -1,66 +1,48 @@
 import torch
 from torch import nn
 from torch.utils.data import Dataset, Subset, DataLoader
+from torchvision import datasets, transforms
 import copy
 from collections import defaultdict
 from random import shuffle
 from typing import List, Tuple
 
-device = "cuda" if torch.cuda.is_available() else "cpu"
+from tinyfl.models.new_fashion_mnist import FashionMNISTModel
+from tinyfl.models.plant_disease_model import (
+    PlantDiseaseModel,
+)
 
 
-class Model(nn.Module):
-    def __init__(self) -> None:
-        super(Model, self).__init__()
-        self.flatten = nn.Flatten()
-        self.layers = nn.Sequential(
-            nn.Linear(28 * 28, 512),
-            nn.ReLU(),
-            nn.Linear(512, 256),
-            nn.ReLU(),
-            nn.Linear(256, 128),
-            nn.ReLU(),
-            nn.Linear(128, 10),
-        )
+models = {
+    "fashion-mnist": FashionMNISTModel,
+    "plant_disease": PlantDiseaseModel,
+}
 
-    def forward(self, x):
-        x = self.flatten(x)
-        x = self.layers(x)
-        return x
+my_datasets = {
+    "fashion-mnist": (
+        datasets.FashionMNIST(
+            root="data",
+            train=True,
+            download=True,
+            transform=transforms.ToTensor(),
+        ),
+        datasets.FashionMNIST(
+            root="data",
+            train=False,
+            download=True,
+            transform=transforms.ToTensor(),
+        ),
+    ),
+    "plant_disease": PlantDiseaseModel.create_datasets(),
+}
 
-
-def create_model() -> Model:
-    model = Model().to(device)
-    return model
-
-
-def train_model(model: Model, epochs: int, trainloader: DataLoader):
-    optimizer = torch.optim.Adam(model.parameters())
-    for _ in range(epochs):
-        model.train()
-        for _, (X, y) in enumerate(trainloader):
-            X, y = X.to(device), y.to(device)
-            pred = model(X)
-            loss = nn.CrossEntropyLoss()(pred, y)
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-
-def test_model(model: Model, testloader: DataLoader) -> Tuple[float, float]:
-    size = len(testloader.dataset)
-    batches = len(testloader)
-    model.eval()
-    test_loss, correct = 0, 0
-    with torch.no_grad():
-        for X, y in testloader:
-            X, y = X.to(device), y.to(device)
-            pred = model(X)
-            test_loss += nn.CrossEntropyLoss()(pred, y).item()
-            correct += (pred.argmax(1) == y).type(torch.float).sum().item()
-    test_loss /= batches
-    correct /= size
-    return 100 * correct, test_loss
+my_loaders = {
+    "fashion-mnist": (
+        DataLoader(my_datasets["fashion-mnist"][0], batch_size=64),
+        DataLoader(my_datasets["fashion-mnist"][1], batch_size=64),
+    ),
+    "plant_disease": PlantDiseaseModel.create_dataloaders(),
+}
 
 
 def fedavg_models(weights):
